pyproject.toml
[tool.poetry]
name = "ernie-plan-and-excute"
version = "0.1.0"
description = ""
authors = ["tobias-hui <ukrabbitessential@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
erniebot-agent = "^0.5.1"
python-dotenv = "^1.0.1"
tavily-python = "^0.3.1"
streamlit = "^1.31.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
